requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# HTTP Client
httpx==0.25.1

# Logging
structlog==23.2.0
python-json-logger==2.0.7

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Document Ingestion & RAG
# Core document processing
docling==1.16.2
docling-core==1.8.1
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
beautifulsoup4==4.12.3
lxml==5.1.0

# Embeddings and ML
sentence-transformers==2.3.1
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
transformers>=4.30.0
huggingface-hub>=0.16.0

# Vector storage
chromadb==0.4.22
chroma-hnswlib>=0.7.3
hnswlib>=0.7.0
posthog>=3.0.0
pydantic-core>=2.10.0

# Text processing
langchain==0.1.4
langchain-text-splitters==0.0.1
langchain-core>=0.1.0
langchain-community>=0.0.1

# Keyword extraction and NLP
keybert==0.8.4
spacy==3.7.2
nltk==3.8.1

# Search
whoosh==2.7.4

# Additional dependencies
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0
pillow>=10.0.0
tqdm>=4.65.0
requests>=2.31.0
typing-extensions>=4.7.0
pyyaml>=6.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
